backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Machine Learning & Data Science
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# OpenAI
openai==1.3.5

# Testing
requests==2.31.0

